price_quote/requirements.txt
alembic==1.13.1
apifairy==1.4.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
dependency-injector==4.41.0
Flask==3.0.1
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
flask-marshmallow==1.1.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.4
marshmallow==3.20.2
marshmallow-sqlalchemy==1.0.0
PyMySQL==1.1.0
python-dotenv==1.0.1
sentry-sdk==1.40.0
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug===3.0.1
pytest==8.1.1
requests==2.31.0
xhtml2pdf==0.2.15
